URL and MySQL Injection Protection

+++ cms/admin/includes/view_all_users.php
@@ -104,11 +104,24 @@
                 }
                 if(isset($_GET['delete'])){
 
-                    $the_user_id = $_GET['delete'];
-                    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-                    $delete_user_query = mysqli_query($connection,$query);
+                    if(isset($_SESSION['user_role'])){
 
-                    header("location: users.php");
+                        if(isset($_SESSION['user_role'])){
+
+                            if($_SESSION['user_role'] == 'admin'){
+
+                                $the_user_id = mysqli_real_escape_string($connection,$_GET['delete']);
+                                $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+                                $delete_user_query = mysqli_query($connection,$query);
+
+                                header("location: users.php");
+                            }
+
+                        }
+
+                    }
+
+
                 }
 
                 ?>
